implemented inheritance POGGERS
