+++ anton_potapov/graph_generation_controller.cpp
@@ -5,25 +5,13 @@
 #include "graph_generation_controller.hpp"
 
 namespace uni_cource_cpp {
-const int MAX_THREADS_COUNT = std::thread::hardware_concurrency();
 
 GraphGenerationController::GraphGenerationController(
     int threads_count,
     int graphs_count,
     GraphGenerator::Params graph_generator_params)
     : graphs_count_(graphs_count), graph_generator_(graph_generator_params) {
-  const auto workers_count = std::min(MAX_THREADS_COUNT, threads_count);
-  for (int i = 0; i < workers_count; ++i) {
-    workers_.emplace_back([this]() -> std::optional<JobCallback> {
-      const std::lock_guard queue_lock(jobs_queue_mutex_);
-      if (!jobs_.empty()) {
-        const auto job = jobs_.front();
-        jobs_.pop();
-        return job;
-      }
-      return std::nullopt;
-    });
-  }
+  init_workers(threads_count);
 }
 
 void GraphGenerationController::generate(
@@ -56,39 +44,6 @@ void GraphGenerationController::generate(
 
   for (auto& worker : workers_) {
     worker.stop();
-  }
-}
-
-GraphGenerationController::Worker::Worker(
-    const GetJobCallback& get_job_callback)
-    : get_job_callback_(get_job_callback) {}
-
-void GraphGenerationController::Worker::start() {
-  assert(state_ != State::Working && "tried to start working worker");
-  state_ = State::Working;
-  thread_ = std::thread([this]() {
-    while (true) {
-      if (state_ == State::ShouldTerminate) {
-        return;
-      }
-      const auto job_optional = get_job_callback_();
-      if (job_optional.has_value()) {
-        const auto& job = job_optional.value();
-        job();
-      }
-    }
-  });
-}
-
-void GraphGenerationController::Worker::stop() {
-  assert(state_ == State::Working && "worker isn't working");
-  state_ = State::ShouldTerminate;
-  thread_.join();
-}
-
-GraphGenerationController::Worker::~Worker() {
-  if (state_ == State::Working) {
-    stop();
   }
 }
 }  // namespace uni_cource_cpp
