+++ anton_potapov/graph_generation_controller.hpp
@@ -5,33 +5,15 @@
 #include <queue>
 #include <thread>
 
+#include "concurrent_controller.hpp"
 #include "graph.hpp"
 #include "graph_generator.hpp"
 
 namespace uni_cource_cpp {
-class GraphGenerationController {
+class GraphGenerationController : public ConcurrentController {
  public:
-  using JobCallback = std::function<void()>;
   using GenStartedCallback = std::function<void(int index)>;
   using GenFinishedCallback = std::function<void(int index, Graph graph)>;
-
-  class Worker {
-   public:
-    using GetJobCallback = std::function<std::optional<JobCallback>()>;
-
-    enum class State { Idle, Working, ShouldTerminate };
-
-    explicit Worker(const GetJobCallback& get_job_callback);
-    ~Worker();
-
-    void start();
-    void stop();
-
-   private:
-    std::thread thread_;
-    GetJobCallback get_job_callback_;
-    std::atomic<State> state_ = State::Idle;
-  };
 
   GraphGenerationController(int threads_count,
                             int graphs_count,
@@ -41,10 +23,6 @@ class GraphGenerationController {
                 const GenFinishedCallback& gen_finished_callback);
 
  private:
-  std::list<Worker> workers_;
-  std::queue<JobCallback> jobs_;
-  std::mutex jobs_queue_mutex_;
-  std::mutex controller_mutex_;
   const int graphs_count_;
   GraphGenerator graph_generator_;
 };
